+++ ObjectPrinting/ISerializingConfig.cs
@@ -1,0 +1,7 @@
+namespace ObjectPrinting
+{
+    interface ISerializingConfig<TOwner>
+    {
+        PrintingConfig<TOwner> SerializingConfig { get; }
+    }
+}
