+++ ObjectPrinting/PrintingConfig.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Globalization;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Runtime.CompilerServices;
@@ -7,57 +6,6 @@ using System.Text;
 
 namespace ObjectPrinting
 {
-    public class SerializingConfig<TOwner, TPropType> : ISerializingConfig<TOwner>
-    {
-        private readonly PrintingConfig<TOwner> _baseClass;
-        
-        public SerializingConfig(PrintingConfig<TOwner> baseClass)
-        {
-            _baseClass = baseClass;
-        }
-        
-        public PrintingConfig<TOwner> Using(Func<TPropType, string> param)
-        {
-            return _baseClass;
-        }
-
-        public PrintingConfig<TOwner> Exclude()
-        {
-            return _baseClass;
-        }
-
-        PrintingConfig<TOwner> ISerializingConfig<TOwner>.SerializingConfig => _baseClass;
-    }
-
-    interface ISerializingConfig<TOwner>
-    {
-        PrintingConfig<TOwner> SerializingConfig { get; }
-    }
-
-    public static class PrintingUsingConfigExtensions
-    {
-        public static PrintingConfig<TOwner> Using<TOwner>(this SerializingConfig<TOwner, int> config, CultureInfo ci)
-        {
-            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
-        }
-
-        public static PrintingConfig<TOwner> Trim<TOwner>(this SerializingConfig<TOwner, string> config)
-        {
-            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
-        }
-        
-        
-        public static PrintingConfig<TOwner> Default<TOwner>(this SerializingConfig<TOwner, int> config)
-        {
-            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
-        }
-        
-        public static PrintingConfig<TOwner> Default<TOwner>(this SerializingConfig<TOwner, double> config)
-        {
-            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
-        }
-    }
-    
     public class PrintingConfig<TOwner>
     {
         public PrintingConfig<TOwner> Exclude<TPorpType>()
