+++ ObjectPrinting/PrintingUsingConfigExtensions.cs
@@ -1,0 +1,28 @@
+using System.Globalization;
+
+namespace ObjectPrinting
+{
+    public static class PrintingUsingConfigExtensions
+    {
+        public static PrintingConfig<TOwner> Using<TOwner>(this SerializingConfig<TOwner, int> config, CultureInfo ci)
+        {
+            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
+        }
+
+        public static PrintingConfig<TOwner> Trim<TOwner>(this SerializingConfig<TOwner, string> config)
+        {
+            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
+        }
+        
+        
+        public static PrintingConfig<TOwner> Default<TOwner>(this SerializingConfig<TOwner, int> config)
+        {
+            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
+        }
+        
+        public static PrintingConfig<TOwner> Default<TOwner>(this SerializingConfig<TOwner, double> config)
+        {
+            return ((ISerializingConfig<TOwner>) config).SerializingConfig;
+        }
+    }
+}
